+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+- Allowed support for Eleventy >1.0.0.
+- Fixed crash occurring when compiling images with no parameters.
+- Fixed images in output of `parse()` having relative URL paths.
+
 ## 1.3.0
 *2021-05-16*
 - Added configuration option `imagesFolder` to configure the folder that images are placed in.